+++ crm-admin/resources/views/admin/booking/index.blade.php
@@ -89,11 +89,11 @@
                     <hr>
 
                     {{-- filter --}}
-
+                    @if(($permissions['booking']->can_edit ?? false))
                     <div class="dt-action-buttons text-end pt-3 pt-md-0">
                         <div class="dt-buttons btn-group flex-wrap">
 
-                            <button class="btn btn-secondary  btn-label-primary me-2" data-bs-toggle="modal"
+                            <!-- <button class="btn btn-secondary  btn-label-primary me-2" data-bs-toggle="modal"
                                 data-bs-target="#basicModal">
                                 <i class="mdi mdi-tray-arrow-down me-sm-1"></i> Import
                             </button>
@@ -103,7 +103,7 @@
                                 aria-controls="DataTables_Table_0" type="button" aria-haspopup="dialog"
                                 aria-expanded="false"><span><i class="mdi mdi-export-variant me-sm-1"></i> <span
                                         class="d-none d-sm-inline-block">Export</span></span></span>
-                            </a>
+                            </a> -->
 
                             <a class="btn btn-secondary btn-primary text-white" href="{{ route('booking.new-trip') }}"
                                 tabindex="0"><span><i class="mdi mdi-plus me-sm-1"></i>
@@ -111,6 +111,7 @@
                             </a>
                         </div>
                     </div>
+                    @endif
                 </div>
                 <div class="card-header">
                     <div class="nav-align-top">
@@ -350,6 +351,10 @@
 
     @section('script')
         <script>
+            var bookingCanView = {{ ($permissions['booking']->can_view ?? false) ? 'true' : 'false' }};
+            var bookingCanEdit = {{ ($permissions['booking']->can_edit ?? false) ? 'true' : 'false' }};
+            var bookingCanDelete = {{ ($permissions['booking']->can_delete ?? false) ? 'true' : 'false' }};
+
             $('th').css('white-space', 'nowrap');
 
             $(document).ready(function() {
@@ -437,11 +442,19 @@
                                 <i class="mdi mdi-dots-vertical"></i>
                                 </button>
                                 <div class="dropdown-menu" style="">
-                                <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>
-                                <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
-                                <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>`;
+                                if (bookingCanView) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
+                                        `;
+                                }
+                                if (bookingCanEdit) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                        `;
+                                }
+                                text += `
                                 </div>
-                                
                             </div>`;
                                 return text;
                             }
@@ -648,8 +661,18 @@
                                 </button>
                                 <div class="dropdown-menu" style="">
                                 <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>
-                                <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
-                                <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                `;
+                                if (bookingCanView) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
+                                        `;
+                                }
+                                if (bookingCanEdit) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                        `;
+                                }
+                                text += `
                                 </div>
                                 
                             </div>`;
@@ -851,8 +874,18 @@
                                 </button>
                                 <div class="dropdown-menu" style="">
                                 <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>
-                                <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
-                                <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                `;
+                                if (bookingCanView) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
+                                        `;
+                                }
+                                if (bookingCanEdit) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                        `;
+                                }
+                                text += `
                                 </div>
                                 
                             </div>`;
@@ -1055,8 +1088,18 @@
                                 </button>
                                 <div class="dropdown-menu" style="">
                                 <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>
-                                <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
-                                <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                `;
+                                if (bookingCanView) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeView}"><i class="mdi mdi-eye-outline me-1"></i> View</a>
+                                        `;
+                                }
+                                if (bookingCanEdit) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                        `;
+                                }
+                                text += `
                                 </div>
                                 
                             </div>`;
@@ -1244,9 +1287,20 @@
                                 </button>
                                 <div class="dropdown-menu" style="">
                                 <a class="dropdown-item waves-effect" href="${routeActivity}"><i class="mdi mdi-book-outline me-1"></i> Activity Log</a>
-                                <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
-                                <a class="dropdown-item waves-effect" data-bs-toggle="modal" onclick="deleteModal('${routeDlt}')"
+                                `;
+                                if (bookingCanEdit) {
+                                    text += `
+                                            <a class="dropdown-item waves-effect" href="${routeEdit}"><i class="mdi mdi-pencil-outline me-1"></i> Edit</a>
+                                        `;
+                                }
+                                if (bookingCanDelete && row.trip_status !== "Completed") {
+                                    text += `
+                                            
+                                            <a class="dropdown-item waves-effect" data-bs-toggle="modal" onclick="deleteModal('${routeDlt}')"
                                             data-bs-target="#deleteModal" href="javaScript:void(0)"><i class="mdi mdi-trash-can-outline me-1"></i> Delete</a>
+                                        `;
+                                }
+                                text += `
                                 </div>
                                 
                             </div>`;
